third_party/blink/web_tests/wpt_internal/overscroll-gestures/parsing/overscroll-anchor-name-valid.html

<!DOCTYPE html>
<meta charset="utf-8">

<title>Overscroll Anchor Name Parsing Test - valid</title>
<meta name="assert" content="overscroll-anchor-name supports the full grammar 'none | <dashed-ident>#'.">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<script>
test_valid_value("overscroll-anchor-name", "none", "none");
test_valid_value("overscroll-anchor-name", "nOnE", "none");
test_valid_value("overscroll-anchor-name", "--foo", "--foo");
test_valid_value("overscroll-anchor-name", "--foo, --bar", "--foo, --bar");
</script>
